+++ src/Enums/EnergyType.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Baspa\EnergyZero\Enums;
+
+enum EnergyType: int
+{
+    case ELECTRICITY = 1;
+    case GAS = 3;
+}